Fix dictionaries deleting

+++ src/main/java/com/shakalyan/trasha/service/DictionaryService.java
@@ -5,11 +5,13 @@ import com.shakalyan.trasha.repository.DictionaryRepo;
 import com.shakalyan.trasha.repository.TranslationRepo;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -29,7 +31,16 @@ public class DictionaryService {
     }
 
     @Transactional
-    public ResponseEntity<String> deleteDictionary(Integer dictionaryId) {
+    public ResponseEntity<String> deleteDictionary(Integer userId, Integer dictionaryId) {
+
+        Optional<Dictionary> dictionary = dictionaryRepo.findById(dictionaryId);
+        if (dictionary.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+        if (!dictionary.get().getUserId().equals(userId)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         translationRepo.deleteAllByDictionaryId(dictionaryId);
         dictionaryRepo.deleteById(dictionaryId);
         return ResponseEntity.ok("Dictionary deleted");
